requirements.txt
# Core Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10
noisereduce>=3.0.0

# Modelling & Tokenization
transformers>=4.35.0
sentencepiece>=0.1.99
pyctcdecode>=0.5.0
kenlm

# Training
accelerate>=0.24.0
wandb>=0.16.0
datasets>=2.14.0
jiwer>=3.0.0

# Serving
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Export & Optimization
onnxruntime>=1.16.0
onnx>=1.15.0

# Utils
pyyaml>=6.0
tqdm>=4.66.0

# HuggingFace Hub (for model sharing)
huggingface_hub>=0.20.0
